requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
PyJWT==2.8.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
email-validator==2.0.0